chore(delete) move Delete component
