+++ next-tavla/src/Admin/scenarios/Boards/components/Column/Actions.tsx
@@ -6,8 +6,8 @@ import { TBoard } from 'types/settings'
 import { useLink } from '../../hooks/useLink'
 import { useToast } from '@entur/alert'
 import classes from './styles.module.css'
-import { DeleteBoardButton as Delete } from '../Delete'
 import { SortableColumn } from './SortableColumn'
+import { DeleteBoardButton as Delete } from './Delete'
 
 function Actions({ board }: { board: TBoard }) {
     const link = useLink(board.id)
