+++ next-tavla/src/Admin/scenarios/Boards/components/Column/Delete.tsx
